requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
matplotlib==3.8.2
pandas==2.1.4
python-dotenv==1.0.0